Added all models of AI!
This is the intial models for the project. They're not that good
for too complex situations but this is what i could in the given time.
There is even still a little room to get more accuracy for this models
but as i said 'no time'.
They solve small puzzles tho!
Since these are individual models, i can spesificly train a certain
puzzle size later on.